Added README files and added file checking to get_state_data and get_country_data.

+++ R/get_country_data.R
@@ -1,5 +1,5 @@
 get_country_data <- function(year = 2024, non_highway = "FALSE"){
-  #' Extracting NBI data into the global environment
+  #' This function extracts NBI country data from the FHWA website into the global environment for use in R.
   #'
   #' @param year The year when the data was collected
   #' @param non_highway Toggle for including non-highway and routes under bridges
@@ -27,8 +27,15 @@ get_country_data <- function(year = 2024, non_highway = "FALSE"){
       website_link <- paste("https://www.fhwa.dot.gov/bridge/nbi/", year, "hwybronefiledel.zip", sep = "")
     }
   }
-  utils::download.file(website_link, destfile = "nbi_data.zip")
-  nbi_data <- readr::read_csv("nbi_data.zip")
-  file.remove("nbi_data.zip")
+  file_name <-substr(website_link, 37, nchar(website_link))
+  if(file.exists(file_name)) {
+    suppressWarnings(nbi_data <- readr::read_csv(file_name))
+  }
+  else {
+    suppressWarnings(utils::download.file(website_link, destfile = "nbi_data.zip"))
+    suppressWarnings(nbi_data <- readr::read_csv("nbi_data.zip"))
+    file.remove("nbi_data.zip")
+  }
+
   return(nbi_data)
 }
